test -1

+++ application/controllers/siswas.php
@@ -21,15 +21,15 @@ class Siswas extends CI_Controller {
         $crud = new grocery_CRUD();
 
         $crud->set_table('tabel_siswa')
-        ->columns('nis','nama','tahun_masuk','jurusan','status','tingkat')
+        ->columns('nisn','nama','tahun_masuk','jurusan','status','tingkat')
         ->set_relation('tahun_masuk','tahun_ajaran','tahun_ajaran',null,'id_tahun_ajaran', 'DESC LIMIT 3')
         ->set_relation('jurusan','tabel_jurusan','nama_jurusan')
         ->unset_fields('id_siswa')
         // ->field_type('tahun_lulus','integer')
         //->field_type('tahun_masuk','integer')
-        ->required_fields('nis','nama_siswa','status', 'jurusan','tahun_masuk')
-        ->unique_fields('nis')
-        // ->set_relation('jns_kelamin','tabel_jenkel','jenis_kelamin')
+        ->required_fields('nisn','nama_siswa','status', 'jurusan','tahun_masuk')
+        ->unique_fields('nisn')
+        // ->set_relation('jns_kelamin','tabel_jenkel','jenisn_kelamin')
         // ->unset_texteditor('alamat','alamat_ortu','alamat_sekolah')
         ->callback_after_insert(array($this,'createUsersiswa'))
         ->callback_before_delete(array($this,'deleteUserSiswa'))
@@ -74,21 +74,77 @@ class Siswas extends CI_Controller {
 
     }
 
+
+
+    public function setBatalLulus()
+    {
+        $siswa = $this->input->post('tidak_lulus');
+
+        $this->siswa->batalKelulusan($siswa);
+        redirect('siswas/menuBatalKelulusan');
+                  
+    }
+
     public function setTidakLulus()
     {
         $siswa = $this->input->post('tidak_lulus');
-        $this->siswa->lulus();
-        if ($siswa != null){
-            foreach ($siswa as $row) {
-                $this->siswa->tidakLulus($row);
-            }
+
+        $this->siswa->kelulusan($siswa);
+        redirect('siswas/menuKelulusan');
+                  
+    }
+
+
+    public function menuBatalKelulusan($value='')
+    {
+        $this->showHeader();
+        $this->load->view('nilai/seleksi_batal_lulus');
+        $this->load->view('footer_general');   
+    }
+
+    public function menuKelulusan($value='')
+    {
+        if ($this->session->userdata('level') != 9){
+            echo "<script>alert('akses ditolak')</script>";
+            // sleep(5);
+            redirect('users/login');
         }
-        echo "<script>alert('Sistem masuk ke tahun ajaran baru!')</script>";
-        redirect('users/home');
+        $this->showHeader();
+        $this->load->view('nilai/menu_kelulusan');
+        $this->load->view('footer_general');   
+    }
+
+    public function menuSKHU($value='')
+    {
+        if ($this->session->userdata('level') != 9){
+            echo "<script>alert('akses ditolak')</script>";
+            // sleep(5);
+            redirect('users/login');
+        }
+        $this->showHeader();
+        $this->load->view('nilai/menu_skhu');
+        $this->load->view('footer_general');
+    }
+
+    public function seleksiKelulusan()
+    {
+        if ($this->session->userdata('level') != 9){
+            echo "<script>alert('akses ditolak')</script>";
+            // sleep(5);
+            redirect('users/login');
+        }
+        $this->showHeader();
+        $this->load->view('nilai/seleksi_kelulusan');
+        $this->load->view('footer_general');  
     }
 
     public function setTunggakan()
     {
+        if ($this->session->userdata('level') != 9){
+            echo "<script>alert('akses ditolak')</script>";
+            // sleep(5);
+            redirect('users/login');
+        }
         $berhasil = true;
         $keMenunggak = $this->input->post('menunggak');
         $keTidakMenunggak = $this->input->post('tidak_menunggak');
@@ -112,6 +168,11 @@ class Siswas extends CI_Controller {
 
     public function gantiTunggakan()
     {
+        if ($this->session->userdata('level') != 9){
+            echo "<script>alert('akses ditolak')</script>";
+            // sleep(5);
+            redirect('users/login');
+        }
         $this->showHeader();
         $this->load->view('siswa/status_tunggakan');    
         $this->load->view('footer_general');  
@@ -173,7 +234,7 @@ class Siswas extends CI_Controller {
 
             $objWorksheet->setCellValue('A1',$string_kartu);
             $objWorksheet->setCellValue('C6',$data_siswa['nama_siswa']);
-            $objWorksheet->setCellValue('C7',$data_siswa['nis']." ");
+            $objWorksheet->setCellValue('C7',$data_siswa['nisn']." ");
             $objWorksheet->setCellValue('C8',$data_kelas['nama_kelas']);
 
             $baris = 11;
@@ -276,6 +337,11 @@ class Siswas extends CI_Controller {
 
     public function setAbsensi($semester)
     {
+        if ($this->session->userdata('level') == 5){
+            echo "<script>alert('akses ditolak')</script>";
+            // sleep(5);
+            redirect('users/login');
+        }
         $berhasil = true;
         print_r($semester);
         print_r($_POST);
@@ -325,8 +391,42 @@ class Siswas extends CI_Controller {
 
     public function cariSiswa($tingkat)
     {
-        $nis = '%'.$this->input->get('search').'%';
-        $query = $this->db->query("SELECT id_siswa as value,nis as text,nama as nama_siswa FROM tabel_siswa WHERE status = '1' AND tingkat='$tingkat' AND nis LIKE '$nis'");
+        $nisn = '%'.$this->input->get('search').'%';
+        $query = $this->db->query("SELECT id_siswa as value,nisn as text,nama as nama_siswa FROM tabel_siswa WHERE status = '1' AND tingkat='$tingkat' AND nisn LIKE '$nisn'");
+        $result = $query->result_array();
+        for ($i=0; $i < sizeof($result); $i++) { 
+            $result[$i]['text'] = $result[$i]['text'] .' - '.$result[$i]['nama_siswa'];
+            $result[$i]['data_siswa'] = null;
+        } 
+
+        // print_r($result);
+        // die();
+        $result = json_encode($result);
+        print_r($result);
+        die();
+    }
+
+    public function cariSiswaLulus()
+    {
+        $nisn = '%'.$this->input->get('search').'%';
+        $query = $this->db->query("SELECT id_siswa as value,nisn as text,nama as nama_siswa FROM tabel_siswa WHERE status = '2' AND tingkat='4' AND nisn LIKE '$nisn'");
+        $result = $query->result_array();
+        for ($i=0; $i < sizeof($result); $i++) { 
+            $result[$i]['text'] = $result[$i]['text'] .' - '.$result[$i]['nama_siswa'];
+            $result[$i]['data_siswa'] = null;
+        } 
+
+        // print_r($result);
+        // die();
+        $result = json_encode($result);
+        print_r($result);
+        die();
+    }
+
+    public function cariSiswaT3()
+    {
+        $nisn = '%'.$this->input->get('search').'%';
+        $query = $this->db->query("SELECT id_siswa as value,nisn as text,nama as nama_siswa FROM tabel_siswa WHERE status = '1' AND tingkat='3' AND nisn LIKE '$nisn'");
         $result = $query->result_array();
         for ($i=0; $i < sizeof($result); $i++) { 
             $result[$i]['text'] = $result[$i]['text'] .' - '.$result[$i]['nama_siswa'];
@@ -342,8 +442,8 @@ class Siswas extends CI_Controller {
 
     public function cariSiswaAll()
     {
-        $nis = '%'.$this->input->get('search').'%';
-        $query = $this->db->query("SELECT id_siswa as value,nis as text,nama as nama_siswa FROM tabel_siswa WHERE status = '1' AND nis LIKE '$nis'");
+        $nisn = '%'.$this->input->get('search').'%';
+        $query = $this->db->query("SELECT id_siswa as value,nisn as text,nama as nama_siswa FROM tabel_siswa WHERE status = '1' AND nisn LIKE '$nisn'");
         $result = $query->result_array();
         for ($i=0; $i < sizeof($result); $i++) { 
             $result[$i]['text'] = $result[$i]['text'] .' - '.$result[$i]['nama_siswa'];
@@ -365,11 +465,32 @@ class Siswas extends CI_Controller {
         $this->load->view('footer_general');
     }
 
+    public function menunggak()
+    {
+        $this->showHeader();
+        $this->load->view('siswa/cegatan_tunggakan');
+        $this->load->view('footer_general');
+    }
+
+    public function belum_waktu()
+    {
+        $this->showHeader();
+        $this->load->view('siswa/cegatan_waktu');
+        $this->load->view('footer_general');
+    }
+
+
+
     function createUsersiswa($post_array, $primary_key)
     {
+        if ($this->session->userdata('level') != 9){
+            echo "<script>alert('akses ditolak')</script>";
+            // sleep(5);
+            redirect('users/login');
+        }
         $credential = $this->db->insert_id();
         $data = array(
-            'user' => $post_array['nis'],
+            'user' => $post_array['nisn'],
             'level' => '4',
             'pass' => sha1('123'),
             'id_transaksi' => $credential
@@ -379,9 +500,14 @@ class Siswas extends CI_Controller {
 
     function deleteUserSiswa($primary_key)
     {
-        $query = $this->db->query("SELECT nis FROM tabel_siswa WHERE id_siswa='$primary_key'");
+        if ($this->session->userdata('level') != 9){
+            echo "<script>alert('akses ditolak')</script>";
+            // sleep(5);
+            redirect('users/login');
+        }
+        $query = $this->db->query("SELECT nisn FROM tabel_siswa WHERE id_siswa='$primary_key'");
         $hasil = $query->first_row();
-        $user = $hasil->nis;
+        $user = $hasil->nisn;
         $this->db->query("DELETE FROM tabel_users WHERE user='$user'");
     }
 }
